Add Pluralize directly to Mist source code
Something about the way it’s set up on Cocoapods means that you can’t
get the Swift 3 compatible version as a dependency in a podspec. So
rather than boil the ocean trying to figure the full solution, we’re
just throwing the code in directly since the library is small and
unlikely to ever need to change.