Remove Record Zone creation code from initializers
We want to create Record Zones when they’re needed, in the same context
where we save Records. This is because a) we don’t need them until that
time, and b) if we try to create them ahead of time, then we have the
potential for multiple write transactions (one creating a Record Zone,
another creating Records that need to go in that zone).